自动更新ETag化

+++ src/AppUI/Vms/AppStatic.cs
@@ -137,27 +137,31 @@ namespace NTMiner.Vms {
             if (!Directory.Exists(updaterDirFullName)) {
                 Directory.CreateDirectory(updaterDirFullName);
             }
-            string ntMinerUpdaterFileFullName = Path.Combine(updaterDirFullName, "NTMinerUpdater.exe");
-            if (!File.Exists(ntMinerUpdaterFileFullName)) {
-                Server.FileUrlService.GetNTMinerUpdaterUrlAsync(downloadFileUrl => {
-                    if (string.IsNullOrEmpty(downloadFileUrl)) {
-                        return;
+            Server.FileUrlService.GetNTMinerUpdaterUrlAsync((downloadFileUrl, headUrl) => {
+                if (string.IsNullOrEmpty(downloadFileUrl) || string.IsNullOrEmpty(headUrl)) {
+                    return;
+                }
+                ETagClient.HeadETagAsync(headUrl, headEtagValue => {
+                    Global.DebugLine($"{headUrl} {headEtagValue}");
+                    string ntMinerUpdaterFileFullName = Path.Combine(updaterDirFullName, "NTMinerUpdater.exe");
+                    string etagKey = "Updater/NTMinerUpdater.exe";
+                    IETag etag = null;
+                    if (!File.Exists(ntMinerUpdaterFileFullName) || !ETagSet.Instance.TryGetETagByKey(etagKey, out etag) || etag.Value != headEtagValue) {
+                        FileDownloader.ShowWindow(downloadFileUrl, "开源矿工更新器", (window, isSuccess, message, saveFileFullName, etagValue) => {
+                            if (isSuccess) {
+                                File.Copy(saveFileFullName, ntMinerUpdaterFileFullName, overwrite: true);
+                                File.Delete(saveFileFullName);
+                                ETagSet.Instance.AddOrUpdateETag(etagKey, etagValue);
+                                window?.Close();
+                                Windows.Cmd.RunClose(ntMinerUpdaterFileFullName, string.Empty);
+                            }
+                        });
                     }
-                    FileDownloader.ShowWindow(downloadFileUrl, "开源矿工更新器", (window, isSuccess, message, saveFileFullName, etagValue) => {
-                        if (isSuccess) {
-                            File.Copy(saveFileFullName, ntMinerUpdaterFileFullName);
-                            File.Delete(saveFileFullName);
-                            string etagKey = "Updater/NTMinerUpdater.exe";
-                            ETagSet.Instance.AddOrUpdateETag(etagKey, etagValue);
-                            window?.Close();
-                            Windows.Cmd.RunClose(ntMinerUpdaterFileFullName, string.Empty);
-                        }
-                    });
+                    else {
+                        Windows.Cmd.RunClose(ntMinerUpdaterFileFullName, string.Empty);
+                    }
                 });
-            }
-            else {
-                Windows.Cmd.RunClose(ntMinerUpdaterFileFullName, string.Empty);
-            }
+            });
         });
         public static ICommand ShowHelp { get; private set; } = new DelegateCommand(() => {
             Process.Start("https://github.com/ntminer/ntminer");
